aider: Fixed error by binding `runGame` function to `gameLogic` object.
# Aider chat conversation:

USER: Error executing code:  this.reset is not a function
Finished processing /execute request

ASSISTANT: The error message indicates that the `this.reset` function is not recognized. This could be due to the context in which `this` is being used. In JavaScript, the value of `this` is determined by how a function is called. It seems like the `this` keyword is not referring to the `gameLogic` object when the `runGame` function is called.

To fix this issue, we need to bind the `runGame` function to the `gameLogic` object. This can be done by modifying the `server.js` file.

Here is the *SEARCH/REPLACE block* for this change:

```javascript
server/server.js
<<<<<<< SEARCH
const { runGame, env } = gameLogic;
=======
const runGame = gameLogic.runGame.bind(gameLogic);
const env = gameLogic.env;
>>>>>>> REPLACE
```

This change ensures that `this` inside the `runGame` function always refers to the `gameLogic` object, regardless of how the function is called. This should fix the error and allow the `reset` function to be called correctly.