aider: The `runGame` function now returns an object with an `index` property and a `logs` property.
# Aider chat conversation:

USER: Error executing code:  Cannot destructure property 'index' of 'runGame(...)' as it is undefined.

ASSISTANT: The error message indicates that the `runGame` function is not returning an object with an `index` property. This could be because the `runGame` function does not return anything at all.

To fix this issue, we need to modify the `runGame` function in the `game_logic.js` file to return an object with the required properties.

Here is the *SEARCH/REPLACE block* for this change:

```javascript
server/game_logic.js
<<<<<<< SEARCH
    // A function to run the game
    runGame: function(code1, code2) {
        // Reset the game
        this.reset();

        // Run the game for a certain number of rounds
        for (let round = 0; round < this.env.numberOfRounds; round++) {
            // Execute the code of the two players
            let move1 = eval(code1);
            let move2 = eval(code2);

            // Record the moves
            this.recordMove(1, move1);
            this.recordMove(2, move2);

            // Update the payoff
            let payoff1 = this.env.gameMatrix[move1][move2][0];
            let payoff2 = this.env.gameMatrix[move1][move2][1];
            this.env.payoff[0] += payoff1;
            this.env.payoff[1] += payoff2;
        }
    }
=======
    // A function to run the game
    runGame: function(code1, code2) {
        // Reset the game
        this.reset();

        // Run the game for a certain number of rounds
        for (let round = 0; round < this.env.numberOfRounds; round++) {
            // Execute the code of the two players
            let move1 = eval(code1);
            let move2 = eval(code2);

            // Record the moves
            this.recordMove(1, move1);
            this.recordMove(2, move2);

            // Update the payoff
            let payoff1 = this.env.gameMatrix[move1][move2][0];
            let payoff2 = this.env.gameMatrix[move1][move2][1];
            this.env.payoff[0] += payoff1;
            this.env.payoff[1] += payoff2;
        }

        // Return the result
        return {
            index: this.env.payoff[0] > this.env.payoff[1] ? 0 : 1,
            logs: this.env.moves
        };
    }
>>>>>>> REPLACE
```

This change ensures that the `runGame` function returns an object with an `index` property (which is 0 if the first player wins and 1 if the second player wins) and a `logs` property (which contains the moves of the players). This should fix the error and allow the `runGame` function to be used correctly.