aider: Define `opponent` as the other player in the `runGame` function in `server/game_logic.js`.
# Aider chat conversation:

USER: ReferenceError: opponent is not defined

ASSISTANT: The error message indicates that the variable `opponent` is not defined in the scope where it's being used. This variable is used in the `if_last_move_cooperate` and `if_last_move_defect` blocks in `client/custom_blocks.js`.

To fix this, we need to define `opponent` in the scope where it's being used. However, it's not clear from the provided code what the value of `opponent` should be. It seems like it should be the ID of the other player, but I don't see where this is being determined.

If you could provide more context about how `opponent` should be defined, I could give a more specific solution.

USER: we could use a similar trick as for `getLastMove` ... we do a let opponent = and defined opponent to be the other player

ASSISTANT: I see, you want to define `opponent` as the other player in the game. In the `runGame` function in `server/game_logic.js`, we can define `opponent` as the other player when we're executing the code for each player. Here's how we can do it:

```javascript
server/game_logic.js
<<<<<<< SEARCH
    // A function to run the game
    runGame: function(code1, code2) {
        // Reset the game
        this.reset();

        // Run the game for a certain number of rounds
        for (let round = 0; round < this.env.numberOfRounds; round++) {
            // Wrap the code in a function before executing
            let move1 = eval('(function() { let getLastMove = ' + this.getLastMove.toString() + '; ' + code1 + '})()');
            let move2 = eval('(function() { let getLastMove = ' + this.getLastMove.toString() + '; ' + code2 + '})()');
=======
    // A function to run the game
    runGame: function(code1, code2) {
        // Reset the game
        this.reset();

        // Run the game for a certain number of rounds
        for (let round = 0; round < this.env.numberOfRounds; round++) {
            // Wrap the code in a function before executing
            let move1 = eval('(function() { let getLastMove = ' + this.getLastMove.toString() + '; let opponent = 1; ' + code1 + '})()');
            let move2 = eval('(function() { let getLastMove = ' + this.getLastMove.toString() + '; let opponent = 0; ' + code2 + '})()');
>>>>>>> REPLACE
```

This change will define `opponent` inside the scope of `eval`, so it can be used in the code being evaluated. Note that this will only work if `code1` and `code2` are using `opponent` as a variable. If they are using it as a property of an object (like `this.opponent`), then we would need to adjust the code accordingly.