aider: Change player count to start from 0 for the first player and 1 for the second player.
# Aider chat conversation:

USER: i am getting a bit nervious about starting to count from 0 sts and from 1 other times, can we consistenlty use 0 for the first player and 1 for the second?

ASSISTANT: Sure, we can make the player count consistent by starting from 0 for the first player and 1 for the second player. This will involve changes in the `server/game_logic.js` and `server/server.js` files.

In `server/game_logic.js`, we need to change the `runGame` function to use 0 and 1 for the players instead of 1 and 2.

In `server/server.js`, we need to change the `app.post('/execute', async (req, res) => {...})` function to use 0 and 1 for the players instead of 1 and 2.

Here are the `SEARCH/REPLACE` blocks for these changes:

```javascript
server/game_logic.js
<<<<<<< SEARCH
            let move1 = eval('(function() { let getLastMove = ' + this.getLastMove.toString() + '; let opponent = 1; ' + code1 + '})()');
            let move2 = eval('(function() { let getLastMove = ' + this.getLastMove.toString() + '; let opponent = 0; ' + code2 + '})()');

            // Record the moves
            this.recordMove(1, move1);
            this.recordMove(2, move2);
=======
            let move0 = eval('(function() { let getLastMove = ' + this.getLastMove.toString() + '; let opponent = 1; ' + code1 + '})()');
            let move1 = eval('(function() { let getLastMove = ' + this.getLastMove.toString() + '; let opponent = 0; ' + code2 + '})()');

            // Record the moves
            this.recordMove(0, move0);
            this.recordMove(1, move1);
>>>>>>> REPLACE
```

```javascript
server/server.js
<<<<<<< SEARCH
    console.log(`Received code from player ${playerId}:\n`, code); // Log the received code

    playerCodes[playerId] = code;
=======
    console.log(`Received code from player ${playerId + 1}:\n`, code); // Log the received code

    playerCodes[playerId + 1] = code;
>>>>>>> REPLACE
```

Please apply these changes and let me know if you encounter any issues.