aider: Replace `move2` with `move1` in the `runGame` function to fix the "ReferenceError: move2 is not defined" error.
# Aider chat conversation:

USER: ok, I feel better now about this; reference error: move2 is not defined

ASSISTANT: The error message "ReferenceError: move2 is not defined" suggests that the variable `move2` is being used in the code but it has not been defined.

Looking at the `runGame` function in `server/game_logic.js`, it seems like `move2` is indeed not defined. The variables `move0` and `move1` are defined to capture the moves of player 0 and player 1 respectively. However, `move2` is used in the calculation of the payoff, which is incorrect.

The payoff should be calculated based on `move0` and `move1`. So, we need to replace `move2` with `move1` in the `runGame` function.

Here is the `SEARCH/REPLACE` block for this change:

```javascript
server/game_logic.js
<<<<<<< SEARCH
            let payoff0 = this.env.gameMatrix[move1][move2][0];
            let payoff1 = this.env.gameMatrix[move1][move2][1];
=======
            let payoff0 = this.env.gameMatrix[move0][move1][0];
            let payoff1 = this.env.gameMatrix[move0][move1][1];
>>>>>>> REPLACE
```

This change should resolve the "ReferenceError: move2 is not defined" error.